requirements.txt
discord.py>=2.3.2
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
aiohttp>=3.8.5
humanize>=4.8.0
PyNaCl>=1.5.0
requests>=2.31.0
yt-dlp>=2023.7.6
spotipy>=2.23.0
google-generativeai>=0.5.0